pyproject.toml
[project]
name = "podcast-ai-agent"
version = "0.1.0"
description = "Download and transcribe audio from YouTube videos"
requires-python = ">=3.10,<3.14"
dependencies = [
    "yt-dlp>=2023.0.0",
    "openai-whisper>=20230314",
    "torch==2.2.2",
    "numpy<2",
    "numba>=0.57.0,<0.60.0",
    "pydub>=0.25.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "textual>=0.47.0",
    "pydantic>=2.12.5",
    "typer>=0.21.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
podcast-ai-agent = "src.__main__:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.uv]
package = true

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
